+++ CHANGELOG.md
@@ -9,11 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
--
+- Editor config for this project.
 
 ### Fixed
 
--
+- Now variables will correctly shared between runners with same namespace.
 
 ### Changed
 
@@ -21,7 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
--
+- Marked current save methods as `Obsolete`.
 
 ## [0.3.1] - 2026-07-09